Updated FileGUI.py to notepad.py Added file.py Modified README.md
