+++ README.md
@@ -1,3 +1,20 @@
 # Programming Languages Semester Project
 
 **Python Text Based Editor**
+
+This project is a simple text-based editor implemented in Python. It consists of two main components:
+
+- **notepad.py**: This file contains the main logic for the text editor, including functions for creating, opening, editing, and saving text files.
+- **file.py**: This file handles file operations such as reading from and writing to files, ensuring that data is correctly managed and stored.
+
+The editor provides a basic interface for text manipulation, making it a useful tool for learning about file handling and user interface design in Python.
+
+## Functional Approach
+
+The project employs a functional programming approach to enhance code readability and maintainability. Key features include:
+
+- **Pure Functions**: Functions are designed to avoid side effects, ensuring that they do not alter any state or data outside their scope.
+- **Immutability**: Data structures are treated as immutable, reducing the likelihood of bugs caused by unintended modifications.
+- **Higher-Order Functions**: Functions that take other functions as arguments or return them as results are used to create more flexible and reusable code.
+
+By leveraging these functional programming principles, the text editor promotes a clean and modular codebase, making it easier to understand and extend.
